Remove unused action types, and split off auth endpoints into another API slice
Commented out tests that broke
