+++ src/IsaacAppTypes.tsx
@@ -34,55 +34,25 @@ import {
 export type Action =
     | {type: ACTION_TYPE.TEST_ACTION}
 
-    | {type: ACTION_TYPE.USER_SNAPSHOT_PARTIAL_UPDATE; userSnapshot: UserSnapshot}
-
     | {type: ACTION_TYPE.ROUTER_PAGE_CHANGE; path: string}
 
     | {type: ACTION_TYPE.API_SERVER_ERROR}
     | {type: ACTION_TYPE.API_GONE_AWAY}
 
-    | {type: ACTION_TYPE.USER_UPDATE_REQUEST}
-    | {type: ACTION_TYPE.USER_UPDATE_RESPONSE_SUCCESS; user: ApiTypes.RegisteredUserDTO}
-    | {type: ACTION_TYPE.USER_UPDATE_RESPONSE_FAILURE}
     | {type: ACTION_TYPE.USER_DETAILS_UPDATE_REQUEST}
     | {type: ACTION_TYPE.USER_DETAILS_UPDATE_RESPONSE_SUCCESS; user: ApiTypes.RegisteredUserDTO}
     | {type: ACTION_TYPE.USER_DETAILS_UPDATE_RESPONSE_FAILURE; errorMessage: string}
-    | {type: ACTION_TYPE.USER_AUTH_SETTINGS_REQUEST}
-    | {type: ACTION_TYPE.USER_AUTH_SETTINGS_RESPONSE_SUCCESS; userAuthSettings: ApiTypes.UserAuthenticationSettingsDTO}
-    | {type: ACTION_TYPE.USER_AUTH_SETTINGS_RESPONSE_FAILURE; errorMessage: string}
-    | {type: ACTION_TYPE.SELECTED_USER_AUTH_SETTINGS_REQUEST}
-    | {type: ACTION_TYPE.SELECTED_USER_AUTH_SETTINGS_RESPONSE_SUCCESS; selectedUserAuthSettings: ApiTypes.UserAuthenticationSettingsDTO}
-    | {type: ACTION_TYPE.SELECTED_USER_AUTH_SETTINGS_RESPONSE_FAILURE; errorMessage: string}
-    | {type: ACTION_TYPE.USER_AUTH_LINK_REQUEST}
-    | {type: ACTION_TYPE.USER_AUTH_LINK_RESPONSE_SUCCESS; provider: AuthenticationProvider; redirectUrl: string}
-    | {type: ACTION_TYPE.USER_AUTH_LINK_RESPONSE_FAILURE; errorMessage: string}
-    | {type: ACTION_TYPE.USER_AUTH_UNLINK_REQUEST}
-    | {type: ACTION_TYPE.USER_AUTH_UNLINK_RESPONSE_SUCCESS; provider: AuthenticationProvider}
-    | {type: ACTION_TYPE.USER_AUTH_UNLINK_RESPONSE_FAILURE; errorMessage: string}
-    | {type: ACTION_TYPE.USER_PREFERENCES_REQUEST}
-    | {type: ACTION_TYPE.USER_PREFERENCES_RESPONSE_SUCCESS; userPreferences: UserPreferencesDTO}
-    | {type: ACTION_TYPE.USER_PREFERENCES_RESPONSE_FAILURE; errorMessage: string}
 
     | {type: ACTION_TYPE.TRANSIENT_USER_CONTEXT_SET_STAGE; stage: STAGE}
     | {type: ACTION_TYPE.TRANSIENT_USER_CONTEXT_SET_EXAM_BOARD; examBoard: EXAM_BOARD}
     | {type: ACTION_TYPE.TRANSIENT_USER_CONTEXT_SET_SHOW_OTHER_CONTENT; showOtherContent: boolean}
 
-    | {type: ACTION_TYPE.USER_LOG_IN_RESPONSE_FAILURE; errorMessage: string}
     | {type: ACTION_TYPE.USER_INCOMING_PASSWORD_RESET_REQUEST}
     | {type: ACTION_TYPE.USER_INCOMING_PASSWORD_RESET_SUCCESS}
     | {type: ACTION_TYPE.USER_INCOMING_PASSWORD_RESET_FAILURE; errorMessage: string}
     | {type: ACTION_TYPE.USER_PASSWORD_RESET_REQUEST}
     | {type: ACTION_TYPE.USER_PASSWORD_RESET_RESPONSE_SUCCESS}
     | {type: ACTION_TYPE.USER_PASSWORD_RESET_RESPONSE_FAILURE; errorMessage: string}
-    | {type: ACTION_TYPE.MY_PROGRESS_REQUEST}
-    | {type: ACTION_TYPE.MY_PROGRESS_RESPONSE_SUCCESS; myProgress: UserProgress}
-    | {type: ACTION_TYPE.MY_PROGRESS_RESPONSE_FAILURE}
-    | {type: ACTION_TYPE.USER_PROGRESS_REQUEST}
-    | {type: ACTION_TYPE.USER_PROGRESS_RESPONSE_SUCCESS; userProgress: UserProgress}
-    | {type: ACTION_TYPE.USER_PROGRESS_RESPONSE_FAILURE}
-    | {type: ACTION_TYPE.USER_SNAPSHOT_REQUEST}
-    | {type: ACTION_TYPE.USER_SNAPSHOT_RESPONSE_SUCCESS; snapshot: UserSnapshot}
-    | {type: ACTION_TYPE.USER_SNAPSHOT_RESPONSE_FAILURE}
     | {type: ACTION_TYPE.AUTHENTICATION_REQUEST_REDIRECT; provider: string}
     | {type: ACTION_TYPE.AUTHENTICATION_REDIRECT; provider: string; redirectUrl: string}
     | {type: ACTION_TYPE.AUTHENTICATION_HANDLE_CALLBACK}
@@ -178,18 +148,6 @@ export type Action =
     | {type: ACTION_TYPE.GROUP_PROGRESS_RESPONSE_SUCCESS; groupId: number; progress: ApiTypes.UserGameboardProgressSummaryDTO[]}
     | {type: ACTION_TYPE.GROUP_PROGRESS_RESPONSE_FAILURE; groupId: number}
 
-    | {type: ACTION_TYPE.CONSTANTS_UNITS_REQUEST}
-    | {type: ACTION_TYPE.CONSTANTS_UNITS_RESPONSE_FAILURE}
-    | {type: ACTION_TYPE.CONSTANTS_UNITS_RESPONSE_SUCCESS; units: string[]}
-
-    | {type: ACTION_TYPE.CONSTANTS_SEGUE_VERSION_REQUEST}
-    | {type: ACTION_TYPE.CONSTANTS_SEGUE_VERSION_RESPONSE_FAILURE}
-    | {type: ACTION_TYPE.CONSTANTS_SEGUE_VERSION_RESPONSE_SUCCESS; segueVersion: string}
-
-    | {type: ACTION_TYPE.CONSTANTS_SEGUE_ENVIRONMENT_REQUEST}
-    | {type: ACTION_TYPE.CONSTANTS_SEGUE_ENVIRONMENT_RESPONSE_FAILURE}
-    | {type: ACTION_TYPE.CONSTANTS_SEGUE_ENVIRONMENT_RESPONSE_SUCCESS; segueEnvironment: string}
-
     | {type: ACTION_TYPE.NOTIFICATIONS_REQUEST}
     | {type: ACTION_TYPE.NOTIFICATIONS_RESPONSE_FAILURE}
     | {type: ACTION_TYPE.NOTIFICATIONS_RESPONSE_SUCCESS; notifications: any[]}
@@ -264,14 +222,6 @@ export type Action =
     | {type: ACTION_TYPE.PROGRESS_REQUEST; assignment: ApiTypes.AssignmentDTO}
     | {type: ACTION_TYPE.PROGRESS_RESPONSE_SUCCESS; assignment: ApiTypes.AssignmentDTO; progress: AppAssignmentProgress[]}
     | {type: ACTION_TYPE.PROGRESS_RESPONSE_FAILURE; assignment: ApiTypes.AssignmentDTO}
-
-    | {type: ACTION_TYPE.CONTENT_VERSION_GET_REQUEST}
-    | {type: ACTION_TYPE.CONTENT_VERSION_GET_RESPONSE_SUCCESS; liveVersion: string}
-    | {type: ACTION_TYPE.CONTENT_VERSION_GET_RESPONSE_FAILURE}
-
-    | {type: ACTION_TYPE.CONTENT_VERSION_SET_REQUEST; version: string}
-    | {type: ACTION_TYPE.CONTENT_VERSION_SET_RESPONSE_SUCCESS; newVersion: string}
-    | {type: ACTION_TYPE.CONTENT_VERSION_SET_RESPONSE_FAILURE}
 
     | {type: ACTION_TYPE.SEARCH_REQUEST; query: string; types: string | undefined}
     | {type: ACTION_TYPE.SEARCH_RESPONSE_SUCCESS; searchResults: ApiTypes.ResultsWrapper<ApiTypes.ContentSummaryDTO>}
