+++ src/app/components/elements/TrustedHtml.tsx
@@ -1,11 +1,10 @@
 import React, {useContext} from "react";
 import {FigureNumberingContext} from "../../../IsaacAppTypes";
-import {AppState} from "../../state/reducers";
 import {useSelector} from "react-redux";
 import {selectors} from "../../state/selectors";
 import {katexify} from "./LaTeX";
 import {useClozeDropRegionsInHtml} from "../content/IsaacClozeQuestion";
-import {isaacApi} from "../../state/slices/api";
+import {authApi} from "../../state/slices/api/auth";
 
 const htmlDom = document.createElement("html");
 function manipulateHtml(html: string) {
@@ -36,7 +35,7 @@ function manipulateHtml(html: string) {
 
 export const TrustedHtml = ({html, span, className}: {html: string; span?: boolean; className?: string}) => {
     const user = useSelector(selectors.user.orNull);
-    const userPreferences = isaacApi.endpoints.userPreferences.useQueryState().currentData;
+    const userPreferences = authApi.endpoints.userPreferences.useQueryState().currentData;
     const booleanNotation = userPreferences?.BOOLEAN_NOTATION || null;
     const screenReaderHoverText = (userPreferences?.BETA_FEATURE && userPreferences?.BETA_FEATURE.SCREENREADER_HOVERTEXT) || false;
