+++ src/app/services/constants.ts
@@ -52,46 +52,13 @@ export const MARKDOWN_RENDERER = new Remarkable({
 export enum ACTION_TYPE {
     TEST_ACTION = "TEST_ACTION",
 
-    USER_SNAPSHOT_PARTIAL_UPDATE = "USER_SNAPSHOT_PARTIAL_UPDATE",
-
     ROUTER_PAGE_CHANGE = "ROUTER_PAGE_CHANGE",
     API_SERVER_ERROR = "API_SERVER_ERROR",
     API_GONE_AWAY = "API_GONE_AWAY",
 
-    USER_LOG_IN_RESPONSE_FAILURE = "USER_LOG_IN_RESPONSE_FAILURE",
-
-    USER_UPDATE_REQUEST = "USER_UPDATE_REQUEST",
-    USER_UPDATE_RESPONSE_SUCCESS = "USER_UPDATE_RESPONSE_SUCCESS",
-    USER_UPDATE_RESPONSE_FAILURE = "USER_UPDATE_RESPONSE_FAILURE",
-
     USER_DETAILS_UPDATE_REQUEST = "USER_DETAILS_UPDATE",
     USER_DETAILS_UPDATE_RESPONSE_SUCCESS = "USER_DETAILS_UPDATE_RESPONSE_SUCCESS",
     USER_DETAILS_UPDATE_RESPONSE_FAILURE = "USER_DETAILS_UPDATE_RESPONSE_FAILURE",
-
-    USER_AUTH_SETTINGS_REQUEST = "USER_AUTH_SETTINGS_REQUEST",
-    USER_AUTH_SETTINGS_RESPONSE_SUCCESS = "USER_AUTH_SETTINGS_RESPONSE_SUCCESS",
-    USER_AUTH_SETTINGS_RESPONSE_FAILURE = "USER_AUTH_SETTINGS_RESPONSE_FAILURE",
-
-    SELECTED_USER_AUTH_SETTINGS_REQUEST = "SELECTED_USER_AUTH_SETTINGS_REQUEST",
-    SELECTED_USER_AUTH_SETTINGS_RESPONSE_SUCCESS = "SELECTED_USER_AUTH_SETTINGS_REQUEST_SUCCESS",
-    SELECTED_USER_AUTH_SETTINGS_RESPONSE_FAILURE = "SELECTED_USER_AUTH_SETTINGS_RESPONSE_FAILURE",
-
-    USER_AUTH_LINK_REQUEST = "USER_AUTH_LINK_REQUEST",
-    USER_AUTH_LINK_RESPONSE_SUCCESS = "USER_AUTH_LINK_RESPONSE_SUCCESS",
-    USER_AUTH_LINK_RESPONSE_FAILURE = "USER_AUTH_LINK_RESPONSE_FAILURE",
-
-    USER_AUTH_UNLINK_REQUEST = "USER_AUTH_UNLINK_REQUEST",
-    USER_AUTH_UNLINK_RESPONSE_SUCCESS = "USER_AUTH_UNLINK_RESPONSE_SUCCESS",
-    USER_AUTH_UNLINK_RESPONSE_FAILURE = "USER_AUTH_UNLINK_RESPONSE_FAILURE",
-
-    USER_AUTH_MFA_CHALLENGE_REQUIRED = "USER_AUTH_MFA_CHALLENGE_REQUIRED",
-    USER_AUTH_MFA_CHALLENGE_REQUEST  = "USER_AUTH_MFA_CHALLENGE_REQUEST",
-    USER_AUTH_MFA_CHALLENGE_SUCCESS = "USER_AUTH_MFA_CHALLENGE_SUCCESS",
-    USER_AUTH_MFA_CHALLENGE_FAILURE = "USER_AUTH_MFA_CHALLENGE_FAILURE",
-
-    USER_PREFERENCES_REQUEST = "USER_PREFERENCES_REQUEST",
-    USER_PREFERENCES_RESPONSE_SUCCESS= "USER_PREFERENCES_RESPONSE_SUCCESS",
-    USER_PREFERENCES_RESPONSE_FAILURE = "USER_PREFERENCES_RESPONSE_FAILURE",
 
     TRANSIENT_USER_CONTEXT_SET_EXAM_BOARD = "TRANSIENT_USER_CONTEXT_SET_EXAM_BOARD",
     TRANSIENT_USER_CONTEXT_SET_STAGE = "TRANSIENT_USER_CONTEXT_SET_STAGE",
@@ -104,18 +71,6 @@ export enum ACTION_TYPE {
     USER_INCOMING_PASSWORD_RESET_REQUEST = "USER_INCOMING_PASSWORD_RESET_REQUEST",
     USER_INCOMING_PASSWORD_RESET_SUCCESS = "USER_INCOMING_PASSWORD_RESET_SUCCESS",
     USER_INCOMING_PASSWORD_RESET_FAILURE = "USER_INCOMING_PASSWORD_RESET_FAILURE",
-
-    MY_PROGRESS_REQUEST = "MY_PROGRESS_REQUEST",
-    MY_PROGRESS_RESPONSE_SUCCESS = "MY_PROGRESS_RESPONSE_SUCCESS",
-    MY_PROGRESS_RESPONSE_FAILURE = "MY_PROGRESS_RESPONSE_FAILURE",
-
-    USER_PROGRESS_REQUEST = "USER_PROGRESS_REQUEST",
-    USER_PROGRESS_RESPONSE_SUCCESS = "USER_PROGRESS_RESPONSE_SUCCESS",
-    USER_PROGRESS_RESPONSE_FAILURE = "USER_PROGRESS_RESPONSE_FAILURE",
-
-    USER_SNAPSHOT_REQUEST = "USER_SNAPSHOT_REQUEST",
-    USER_SNAPSHOT_RESPONSE_SUCCESS = "USER_SNAPSHOT_RESPONSE_SUCCESS",
-    USER_SNAPSHOT_RESPONSE_FAILURE = "USER_SNAPSHOT_RESPONSE_FAILURE",
 
     AUTHENTICATION_REQUEST_REDIRECT = "AUTHENTICATION_REQUEST_REDIRECT",
     AUTHENTICATION_REDIRECT = "AUTHENTICATION_REDIRECT",
@@ -203,18 +158,6 @@ export enum ACTION_TYPE {
     GROUP_PROGRESS_REQUEST = "GROUP_PROGRESS_REQUEST",
     GROUP_PROGRESS_RESPONSE_SUCCESS = "GROUP_PROGRESS_RESPONSE_SUCCESS",
     GROUP_PROGRESS_RESPONSE_FAILURE = "GROUP_PROGRESS_RESPONSE_FAILURE",
-
-    CONSTANTS_UNITS_REQUEST = "CONSTANTS_UNITS_REQUEST",
-    CONSTANTS_UNITS_RESPONSE_SUCCESS = "CONSTANTS_UNITS_SUCCESS",
-    CONSTANTS_UNITS_RESPONSE_FAILURE = "CONSTANTS_UNITS_RESPONSE_FAILURE",
-
-    CONSTANTS_SEGUE_VERSION_REQUEST = "CONSTANTS_SEGUE_VERSION_REQUEST",
-    CONSTANTS_SEGUE_VERSION_RESPONSE_SUCCESS = "CONSTANTS_SEGUE_VERSION_RESPONSE_SUCCESS",
-    CONSTANTS_SEGUE_VERSION_RESPONSE_FAILURE = "CONSTANTS_SEGUE_VERSION_RESPONSE_FAILURE",
-
-    CONSTANTS_SEGUE_ENVIRONMENT_REQUEST = "CONSTANTS_SEGUE_ENVIRONMENT_REQUEST",
-    CONSTANTS_SEGUE_ENVIRONMENT_RESPONSE_SUCCESS = "CONSTANTS_SEGUE_ENVIRONMENT_RESPONSE_SUCCESS",
-    CONSTANTS_SEGUE_ENVIRONMENT_RESPONSE_FAILURE = "CONSTANTS_SEGUE_ENVIRONMENT_RESPONSE_FAILURE",
 
     NOTIFICATIONS_REQUEST = "NOTIFICATIONS_REQUEST",
     NOTIFICATIONS_RESPONSE_SUCCESS = "NOTIFICATIONS_RESPONSE_SUCCESS",
@@ -380,14 +323,6 @@ export enum ACTION_TYPE {
     PROGRESS_REQUEST = "PROGRESS_REQUEST",
     PROGRESS_RESPONSE_SUCCESS = "PROGRESS_SUCCESS",
     PROGRESS_RESPONSE_FAILURE = "PROGRESS_RESPONSE_FAILURE",
-
-    CONTENT_VERSION_GET_REQUEST = "CONTENT_VERSION_GET_REQUEST",
-    CONTENT_VERSION_GET_RESPONSE_SUCCESS = "CONTENT_VERSION_GET_RESPONSE_SUCCESS",
-    CONTENT_VERSION_GET_RESPONSE_FAILURE = "CONTENT_VERSION_GET_RESPONSE_FAILURE",
-
-    CONTENT_VERSION_SET_REQUEST = "CONTENT_VERSION_SET_REQUEST",
-    CONTENT_VERSION_SET_RESPONSE_SUCCESS = "CONTENT_VERSION_SET_RESPONSE_SUCCESS",
-    CONTENT_VERSION_SET_RESPONSE_FAILURE = "CONTENT_VERSION_SET_RESPONSE_FAILURE",
 
     SEARCH_REQUEST = "SEARCH_REQUEST",
     SEARCH_RESPONSE_SUCCESS = "SEARCH_RESPONSE_SUCCESS",
