+++ src/app/state/reducers/staticState.ts
@@ -11,17 +11,3 @@ export const news = (news: NewsState = null, action: Action) => {
             return news;
     }
 };
-
-type ConstantsState = {units?: string[]; segueVersion?: string; segueEnvironment?: string} | null;
-export const constants = (constants: ConstantsState = null, action: Action) => {
-    switch (action.type) {
-        case ACTION_TYPE.CONSTANTS_UNITS_RESPONSE_SUCCESS:
-            return {...constants, units: action.units};
-        case ACTION_TYPE.CONSTANTS_SEGUE_VERSION_RESPONSE_SUCCESS:
-            return {...constants, segueVersion: action.segueVersion};
-        case ACTION_TYPE.CONSTANTS_SEGUE_ENVIRONMENT_RESPONSE_SUCCESS:
-            return {...constants, segueEnvironment: action.segueEnvironment};
-        default:
-            return constants;
-    }
-};
